rearrange code to not break appservice kube
